requirements.txt
# Flask and Web UI
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.2
Flask-WTF==1.1.1
python-dotenv==1.0.0

# Database
pymongo==4.5.0

# YouTube API
google-api-python-client==2.97.0
youtube-transcript-api==0.6.1

# Task Queue and Processing
celery==5.3.1
kombu==5.3.1
billiard==4.1.0

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-celery==3.3.0

# Email
pydantic==2.0.3

# AI/NLP
openai==0.28.1
tiktoken==0.5.1
langchain==0.0.286
langchain-openai==0.0.2.post1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
tenacity==8.2.3

# Core dependencies
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
gunicorn==20.1.0

# Message broker
pika==1.3.1

# Task processing
flower==1.2.0

# Storage
motor==3.1.1

# LinkedIn API
linkedin-api==2.0.0
requests-oauthlib==1.3.1

# AI/ML
transformers==4.27.4
nltk==3.8.1
spacy==3.5.1

# UI
bootstrap-flask==2.2.0

# Development & testing
pytest==7.3.1
pytest-cov==4.1.0
black==23.3.0
isort==5.12.0
flake8==6.0.0 